[FOLD] More splitting
