+++ Builds/VisualStudio2013/RippleD.vcxproj.filters
@@ -4086,6 +4086,9 @@
     <ClInclude Include="..\..\src\ripple\shamap\TreeNodeCache.h">
       <Filter>ripple\shamap</Filter>
     </ClInclude>
+    <ClInclude Include="..\..\src\ripple\test\jtx.h">
+      <Filter>ripple\test</Filter>
+    </ClInclude>
     <ClInclude Include="..\..\src\ripple\test\jtx\Account.h">
       <Filter>ripple\test\jtx</Filter>
     </ClInclude>
@@ -4096,6 +4099,9 @@
       <Filter>ripple\test\jtx</Filter>
     </ClInclude>
     <ClInclude Include="..\..\src\ripple\test\jtx\Env.h">
+      <Filter>ripple\test\jtx</Filter>
+    </ClInclude>
+    <ClInclude Include="..\..\src\ripple\test\jtx\fee.h">
       <Filter>ripple\test\jtx</Filter>
     </ClInclude>
     <ClCompile Include="..\..\src\ripple\test\jtx\impl\Account.cpp">
@@ -4110,6 +4116,9 @@
     <ClCompile Include="..\..\src\ripple\test\jtx\impl\Env_test.cpp">
       <Filter>ripple\test\jtx\impl</Filter>
     </ClCompile>
+    <ClCompile Include="..\..\src\ripple\test\jtx\impl\fee.cpp">
+      <Filter>ripple\test\jtx\impl</Filter>
+    </ClCompile>
     <ClCompile Include="..\..\src\ripple\test\jtx\impl\json.cpp">
       <Filter>ripple\test\jtx\impl</Filter>
     </ClCompile>
@@ -4122,7 +4131,22 @@
     <ClCompile Include="..\..\src\ripple\test\jtx\impl\owners.cpp">
       <Filter>ripple\test\jtx\impl</Filter>
     </ClCompile>
+    <ClCompile Include="..\..\src\ripple\test\jtx\impl\paths.cpp">
+      <Filter>ripple\test\jtx\impl</Filter>
+    </ClCompile>
+    <ClCompile Include="..\..\src\ripple\test\jtx\impl\sendmax.cpp">
+      <Filter>ripple\test\jtx\impl</Filter>
+    </ClCompile>
+    <ClCompile Include="..\..\src\ripple\test\jtx\impl\seq.cpp">
+      <Filter>ripple\test\jtx\impl</Filter>
+    </ClCompile>
+    <ClCompile Include="..\..\src\ripple\test\jtx\impl\sig.cpp">
+      <Filter>ripple\test\jtx\impl</Filter>
+    </ClCompile>
     <ClCompile Include="..\..\src\ripple\test\jtx\impl\ticket.cpp">
+      <Filter>ripple\test\jtx\impl</Filter>
+    </ClCompile>
+    <ClCompile Include="..\..\src\ripple\test\jtx\impl\txflags.cpp">
       <Filter>ripple\test\jtx\impl</Filter>
     </ClCompile>
     <ClInclude Include="..\..\src\ripple\test\jtx\json.h">
@@ -4137,10 +4161,28 @@
     <ClInclude Include="..\..\src\ripple\test\jtx\owners.h">
       <Filter>ripple\test\jtx</Filter>
     </ClInclude>
+    <ClInclude Include="..\..\src\ripple\test\jtx\paths.h">
+      <Filter>ripple\test\jtx</Filter>
+    </ClInclude>
+    <ClInclude Include="..\..\src\ripple\test\jtx\sendmax.h">
+      <Filter>ripple\test\jtx</Filter>
+    </ClInclude>
+    <ClInclude Include="..\..\src\ripple\test\jtx\seq.h">
+      <Filter>ripple\test\jtx</Filter>
+    </ClInclude>
+    <ClInclude Include="..\..\src\ripple\test\jtx\sig.h">
+      <Filter>ripple\test\jtx</Filter>
+    </ClInclude>
     <ClInclude Include="..\..\src\ripple\test\jtx\tags.h">
       <Filter>ripple\test\jtx</Filter>
     </ClInclude>
+    <ClInclude Include="..\..\src\ripple\test\jtx\ter.h">
+      <Filter>ripple\test\jtx</Filter>
+    </ClInclude>
     <ClInclude Include="..\..\src\ripple\test\jtx\ticket.h">
+      <Filter>ripple\test\jtx</Filter>
+    </ClInclude>
+    <ClInclude Include="..\..\src\ripple\test\jtx\txflags.h">
       <Filter>ripple\test\jtx</Filter>
     </ClInclude>
     <ClCompile Include="..\..\src\ripple\unity\app_ledger.cpp">
