+++ src/ripple/test/jtx/JTx.h
@@ -132,187 +132,6 @@ parse (Json::Value const& jv);
 
 //------------------------------------------------------------------------------
 //
-// Funclets
-//
-//------------------------------------------------------------------------------
-
-/** Set the fee on a JTx. */
-class fee
-{
-private:
-    STAmount v_;
-    boost::tribool b_ =
-        boost::logic::indeterminate;
-
-public:
-    explicit
-    fee (autofill_t)
-        : b_(true)
-    {
-    }
-
-    explicit
-    fee (none_t)
-        : b_(false)
-    {
-    }
-
-    explicit
-    fee (STAmount const& v)
-        : v_(v)
-    {
-        if (! isXRP(v_))
-            throw std::runtime_error(
-                "fee: not XRP");
-    }
-
-    void
-    operator()(Env const&, JTx& jt) const;
-};
-
-/** Set Paths, SendMax on a JTx. */
-class paths
-{
-private:
-    Issue in_;
-    int depth_;
-    unsigned int limit_;
-
-public:
-    paths (Issue const& in,
-            int depth = 7, unsigned int limit = 4)
-        : in_(in)
-        , depth_(depth)
-        , limit_(limit)
-    {
-    }
-
-    void
-    operator()(Env const&, JTx& jtx) const;
-};
-
-
-/** Sets the SendMax on a JTx. */
-class sendmax
-{
-private:
-    STAmount amount_;
-
-public:
-    sendmax (STAmount const& amount)
-        : amount_(amount)
-    {
-    }
-
-    void
-    operator()(Env const&, JTx& jtx) const;
-};
-
-/** Set the flags on a JTx. */
-class txflags
-{
-private:
-    std::uint32_t v_;
-
-public:
-    explicit
-    txflags (std::uint32_t v)
-        : v_(v)
-    {
-    }
-
-    void
-    operator()(Env const&, JTx& jt) const;
-};
-
-/** Set the sequence number on a JTx. */
-struct seq
-{
-private:
-    std::uint32_t v_;
-    boost::tribool b_ =
-        boost::logic::indeterminate;
-
-public:
-    explicit
-    seq (autofill_t)
-        : b_(true)
-    {
-    }
-
-    explicit
-    seq (none_t)
-        : b_(false)
-    {
-    }
-
-    explicit
-    seq (std::uint32_t v)
-        : v_(v)
-    {
-    }
-
-    void
-    operator()(Env const&, JTx& jt) const;
-};
-
-/** Set the regular signature on a JTx.
-    @note For multisign, use msig.
-*/
-class sig
-{
-private:
-    Account account_;
-    boost::tribool b_ =
-        boost::logic::indeterminate;
-
-public:
-    explicit
-    sig (autofill_t)
-        : b_(true)
-    {
-    }
-
-    explicit
-    sig (none_t)
-        : b_(false)
-    {
-    }
-
-    explicit
-    sig (Account const& account)
-        : account_(account)
-    {
-    }
-
-    void
-    operator()(Env const&, JTx& jt) const;
-};
-
-/** Set the expected result code for a JTx
-    The test will fail if the code doesn't match.
-*/
-class ter
-{
-private:
-    TER v_;
-
-public:
-    explicit
-    ter (TER v)
-        : v_(v)
-    {
-    }
-
-    void
-    operator()(Env const&, JTx& jt) const
-    {
-        jt.ter = v_;
-    }
-};
-
-//------------------------------------------------------------------------------
-//
 // Conditions
 //
 //------------------------------------------------------------------------------
