+++ src/ripple/test/jtx/fee.h
@@ -17,29 +17,54 @@
 */
 //==============================================================================
 
-#ifndef RIPPLE_TEST_JTX_H_INCLUDED
-#define RIPPLE_TEST_JTX_H_INCLUDED
+#ifndef RIPPLE_TEST_JTX_FEE_H_INCLUDED
+#define RIPPLE_TEST_JTX_FEE_H_INCLUDED
 
-// Convenience header that includes everything
-
-#include <ripple/test/jtx/Account.h>
-#include <ripple/test/jtx/amounts.h>
-#include <ripple/test/jtx/any.h>
-//#include <ripple/test/jtx/balance.h>
 #include <ripple/test/jtx/Env.h>
-#include <ripple/test/jtx/fee.h>
-#include <ripple/test/jtx/json.h>
-#include <ripple/test/jtx/JTx.h>
-#include <ripple/test/jtx/multisign.h>
-#include <ripple/test/jtx/owners.h>
-#include <ripple/test/jtx/paths.h>
-#include <ripple/test/jtx/sendmax.h>
-#include <ripple/test/jtx/seq.h>
-#include <ripple/test/jtx/sig.h>
 #include <ripple/test/jtx/tags.h>
-#include <ripple/test/jtx/ter.h>
-#include <ripple/test/jtx/ticket.h>
-#include <ripple/test/jtx/txflags.h>
+#include <ripple/protocol/STAmount.h>
+#include <boost/logic/tribool.hpp>
+
+namespace ripple {
+namespace test {
+namespace jtx {
+
+/** Set the fee on a JTx. */
+class fee
+{
+private:
+    STAmount v_;
+    boost::tribool b_ =
+        boost::logic::indeterminate;
+
+public:
+    explicit
+    fee (autofill_t)
+        : b_(true)
+    {
+    }
+
+    explicit
+    fee (none_t)
+        : b_(false)
+    {
+    }
+
+    explicit
+    fee (STAmount const& v)
+        : v_(v)
+    {
+        if (! isXRP(v_))
+            throw std::runtime_error(
+                "fee: not XRP");
+    }
+
+    void
+    operator()(Env const&, JTx& jt) const;
+};
+
+} // jtx
+} // test
+} // ripple
 
 #endif
-
