+++ src/ripple/test/jtx/impl/Env.cpp
@@ -19,6 +19,9 @@
 
 #include <BeastConfig.h>
 #include <ripple/test/jtx/Env.h>
+#include <ripple/test/jtx/fee.h>
+#include <ripple/test/jtx/seq.h>
+#include <ripple/test/jtx/sig.h>
 #include <ripple/app/paths/FindPaths.h>
 #include <ripple/app/tx/TransactionEngine.h>
 #include <ripple/basics/Slice.h>
