+++ src/ripple/test/jtx/impl/JTx.cpp
@@ -83,79 +83,6 @@ parse (Json::Value const& jv)
     return std::move(*p.object);
 }
 
-void
-fee::operator()(Env const&, JTx& jt) const
-{
-    if (boost::indeterminate(b_))
-        jt[jss::Fee] =
-            v_.getJson(0);
-    else
-        jt.fill_fee = b_;
-}
-
-void
-paths::operator()(Env const& env, JTx& jt) const
-{
-    auto& jv = jt.jv;
-    auto const from = env.lookup(
-        jv[jss::Account].asString());
-    auto const to = env.lookup(
-        jv[jss::Destination].asString());
-    auto const amount = amountFromJson(
-        sfAmount, jv[jss::Amount]);
-    STPath fp;
-    STPathSet ps;
-    auto const found = findPathsForOneIssuer(
-        std::make_shared<RippleLineCache>(
-            env.ledger), from, to,
-                in_, amount,
-                    depth_, limit_, ps, fp);
-    // VFALCO TODO API to allow caller to examine the STPathSet
-    // VFALCO isDefault should be renamed to empty()
-    if (found && ! ps.isDefault())
-        jv[jss::Paths] = ps.getJson(0);
-}
-
-void
-sendmax::operator()(Env const& env, JTx& jt) const
-{
-    jt.jv[jss::SendMax] = amount_.getJson(0);
-}
-
-void
-txflags::operator()(Env const&, JTx& jt) const
-{
-    jt[jss::Flags] =
-        v_ /*| tfUniversal*/;
-}
-
-void
-seq::operator()(Env const&, JTx& jt) const
-{
-    if (boost::indeterminate(b_))
-        jt[jss::Sequence] = v_;
-    else
-        jt.fill_seq = b_;
-}
-
-void
-sig::operator()(Env const&, JTx& jt) const
-{
-    if(boost::indeterminate(b_))
-    {
-        // VFALCO Inefficient pre-C++14
-        auto const account = account_;
-        jt.signer = [account](Env&, JTx& jt)
-        {
-            jtx::sign(jt.jv, account);
-        };
-    }
-    else
-    {
-        jt.fill_sig = b_;
-    }
-}
-
 //------------------------------------------------------------------------------
 //
 // Conditions
