+++ src/ripple/test/jtx/sendmax.h
@@ -17,29 +17,34 @@
 */
 //==============================================================================
 
-#ifndef RIPPLE_TEST_JTX_H_INCLUDED
-#define RIPPLE_TEST_JTX_H_INCLUDED
+#ifndef RIPPLE_TEST_JTX_SENDMAX_H_INCLUDED
+#define RIPPLE_TEST_JTX_SENDMAX_H_INCLUDED
 
-// Convenience header that includes everything
-
-#include <ripple/test/jtx/Account.h>
-#include <ripple/test/jtx/amounts.h>
-#include <ripple/test/jtx/any.h>
-//#include <ripple/test/jtx/balance.h>
 #include <ripple/test/jtx/Env.h>
-#include <ripple/test/jtx/fee.h>
-#include <ripple/test/jtx/json.h>
-#include <ripple/test/jtx/JTx.h>
-#include <ripple/test/jtx/multisign.h>
-#include <ripple/test/jtx/owners.h>
-#include <ripple/test/jtx/paths.h>
-#include <ripple/test/jtx/sendmax.h>
-#include <ripple/test/jtx/seq.h>
-#include <ripple/test/jtx/sig.h>
-#include <ripple/test/jtx/tags.h>
-#include <ripple/test/jtx/ter.h>
-#include <ripple/test/jtx/ticket.h>
-#include <ripple/test/jtx/txflags.h>
+#include <ripple/protocol/STAmount.h>
+
+namespace ripple {
+namespace test {
+namespace jtx {
+
+/** Sets the SendMax on a JTx. */
+class sendmax
+{
+private:
+    STAmount amount_;
+
+public:
+    sendmax (STAmount const& amount)
+        : amount_(amount)
+    {
+    }
+
+    void
+    operator()(Env const&, JTx& jtx) const;
+};
+
+} // jtx
+} // test
+} // ripple
 
 #endif
-
