+++ src/ripple/test/jtx/ter.h
@@ -17,29 +17,39 @@
 */
 //==============================================================================
 
-#ifndef RIPPLE_TEST_JTX_H_INCLUDED
-#define RIPPLE_TEST_JTX_H_INCLUDED
+#ifndef RIPPLE_TEST_JTX_TER_H_INCLUDED
+#define RIPPLE_TEST_JTX_TER_H_INCLUDED
 
-// Convenience header that includes everything
-
-#include <ripple/test/jtx/Account.h>
-#include <ripple/test/jtx/amounts.h>
-#include <ripple/test/jtx/any.h>
-//#include <ripple/test/jtx/balance.h>
 #include <ripple/test/jtx/Env.h>
-#include <ripple/test/jtx/fee.h>
-#include <ripple/test/jtx/json.h>
-#include <ripple/test/jtx/JTx.h>
-#include <ripple/test/jtx/multisign.h>
-#include <ripple/test/jtx/owners.h>
-#include <ripple/test/jtx/paths.h>
-#include <ripple/test/jtx/sendmax.h>
-#include <ripple/test/jtx/seq.h>
-#include <ripple/test/jtx/sig.h>
-#include <ripple/test/jtx/tags.h>
-#include <ripple/test/jtx/ter.h>
-#include <ripple/test/jtx/ticket.h>
-#include <ripple/test/jtx/txflags.h>
+
+namespace ripple {
+namespace test {
+namespace jtx {
+
+/** Set the expected result code for a JTx
+    The test will fail if the code doesn't match.
+*/
+class ter
+{
+private:
+    TER v_;
+
+public:
+    explicit
+    ter (TER v)
+        : v_(v)
+    {
+    }
+
+    void
+    operator()(Env const&, JTx& jt) const
+    {
+        jt.ter = v_;
+    }
+};
+
+} // jtx
+} // test
+} // ripple
 
 #endif
-
